Seperate the dart and the flutter parts of this sdk (#434)
* Set TODO's

Set TODO's, needed for https://github.com/parse-community/Parse-SDK-Flutter/issues/78#issuecomment-682154173

* First step of seperating the flutter and the dart parts

* added one todo

* remove connectivity from the dart part

* Removed path_provider from the dart part

Not so happy with the fact that `CoreStoreSembastImp implements sdk.CoreStoreSembastImp` in `parse_server_sdk.dart` but I didn't find a better solution.

* Removed shared_preferences from the dart part

* Update pubspec.yaml

* Update parse_server_sdk_dart.dart

* removed dart:ui from dart part

* removed flutter from dart part

* fixed tests
