+++ lib/src/network/parse_live_query.dart
@@ -33,15 +33,20 @@ class Subscription<T extends ParseObject> {
 
 enum LiveQueryClientEvent { CONNECTED, DISCONNECTED, USER_DISCONNECTED }
 
-class LiveQueryReconnectingController with WidgetsBindingObserver {
+class LiveQueryReconnectingController {
   LiveQueryReconnectingController(
-      this._reconnect, this._eventStream, this.debug) {
-    //Connectivity works differently on web
-    if (!parseIsWeb) {
-      Connectivity().checkConnectivity().then(_connectivityChanged);
-      Connectivity().onConnectivityChanged.listen(_connectivityChanged);
+    this._reconnect,
+    this._eventStream,
+    this.debug,
+  ) {
+    final ParseConnectivityProvider connectivityProvider =
+        ParseCoreData().connectivityProvider;
+    if (connectivityProvider != null) {
+      connectivityProvider.checkConnectivity().then(_connectivityChanged);
+      connectivityProvider.connectivityStream.listen(_connectivityChanged);
     } else {
-      _connectivityChanged(ConnectivityResult.wifi);
+      print(
+          'LiveQuery does not work, if there is ParseConnectivityProvider provided.');
     }
     _eventStream.listen((LiveQueryClientEvent event) {
       switch (event) {
@@ -67,7 +72,7 @@ class LiveQueryReconnectingController with WidgetsBindingObserver {
         print('$DEBUG_TAG: $event');
       }
     });
-    WidgetsBinding.instance.addObserver(this);
+    ParseCoreData().appResumedStream?.listen((void _) => _setReconnect());
   }
 
   static List<int> get retryInterval => ParseCoreData().liveListRetryIntervals;
@@ -84,26 +89,15 @@ class LiveQueryReconnectingController with WidgetsBindingObserver {
 
   Timer _currentTimer;
 
-  void _connectivityChanged(ConnectivityResult state) {
-    if (!_isOnline && state != ConnectivityResult.none) {
+  void _connectivityChanged(ParseConnectivityResult state) {
+    if (!_isOnline && state != ParseConnectivityResult.none) {
       _retryState = 0;
     }
-    _isOnline = state != ConnectivityResult.none;
+    _isOnline = state != ParseConnectivityResult.none;
     if (debug) {
       print('$DEBUG_TAG: $state');
     }
     _setReconnect();
-  }
-
-  @override
-  void didChangeAppLifecycleState(AppLifecycleState state) {
-    switch (state) {
-      case AppLifecycleState.resumed:
-        _setReconnect();
-        break;
-      default:
-        break;
-    }
   }
 
   void _setReconnect() {
