+++ pubspec.yaml
@@ -13,17 +13,17 @@ dependencies:
 
   # Networking
   web_socket_channel: ^1.1.0
-  connectivity: ^0.4.9+2
+  connectivity: ^0.4.9+2 # only used in the flutter part
 
   #Database
   sembast: ^2.4.7+6
   xxtea: ^2.0.3
-  shared_preferences: ^0.5.10
+  shared_preferences: ^0.5.10 # only used in the flutter part
 
   # Utils
-  path_provider: ^1.6.14
+  path_provider: ^1.6.14 # only used in the flutter part
   uuid: ^2.2.2
-  package_info: ^0.4.3
+  package_info: ^0.4.3 # only used in the flutter part
   meta: ^1.1.8
   path: ^1.7.0
