+++ test/parse_client_configuration_test.dart
@@ -10,8 +10,11 @@ void main() {
         clientKey: 'clientKey',
         liveQueryUrl: 'liveQueryUrl',
         appName: 'appName',
+        appPackageName: 'somePackageName',
+        appVersion: 'someAppVersion',
         masterKey: 'masterKey',
         sessionId: 'sessionId',
+        fileDirectory: 'someDirectory',
         debug: true);
 
     expect(ParseCoreData().applicationId, 'appId');
@@ -19,8 +22,11 @@ void main() {
     expect(ParseCoreData().clientKey, 'clientKey');
     expect(ParseCoreData().liveQueryURL, 'liveQueryUrl');
     expect(ParseCoreData().appName, 'appName');
+    expect(ParseCoreData().appPackageName, 'somePackageName');
+    expect(ParseCoreData().appVersion, 'someAppVersion');
     expect(ParseCoreData().masterKey, 'masterKey');
     expect(ParseCoreData().sessionId, 'sessionId');
     expect(ParseCoreData().debug, true);
+    expect(ParseCoreData().fileDirectory, 'someDirectory');
   });
 }
