+++ test/parse_query_test.dart
@@ -12,10 +12,22 @@ void main() {
     test('whereRelatedTo', () async {
       final MockClient client = MockClient();
 
-      await Parse().initialize('appId', 'https://test.parse.com', debug: true);
+      await Parse().initialize(
+        'appId',
+        'https://test.parse.com',
+        debug: true,
+        // to prevent automatic detection
+        fileDirectory: 'someDirectory',
+        // to prevent automatic detection
+        appName: 'appName',
+        // to prevent automatic detection
+        appPackageName: 'somePackageName',
+        // to prevent automatic detection
+        appVersion: 'someAppVersion',
+      );
 
       final QueryBuilder<ParseObject> queryBuilder =
-      QueryBuilder<ParseObject>(ParseObject('_User', client: client));
+          QueryBuilder<ParseObject>(ParseObject('_User', client: client));
       queryBuilder.whereRelatedTo('likes', 'Post', '8TOXdXf3tz');
 
       when(client.data).thenReturn(ParseCoreData());
@@ -27,7 +39,7 @@ void main() {
 
       final Uri expectedQuery = Uri(
           query:
-          'where={"\$relatedTo":{"object":{"__type":"Pointer","className":"Post","objectId":"8TOXdXf3tz"},"key":"likes"}}');
+              'where={"\$relatedTo":{"object":{"__type":"Pointer","className":"Post","objectId":"8TOXdXf3tz"},"key":"likes"}}');
       expect(result.query, expectedQuery.query);
     });
   });
